Update Wildfire
This commit updates Wildfire to League S0W5
