+++ RLBotPack/Wildfire_League_S0W5/python/wildfire.cfg
@@ -1,0 +1,31 @@
+[Locations]
+# Path to loadout config from runner
+looks_config = wildfireAppearance.cfg
+
+# Path to module from runner
+python_file = wildfire.py
+
+logo_file = logo.png
+
+# Name that will be displayed in game
+name = Wildfire
+
+[Bot Parameters]
+java_executable_path = ../Wildfire/bin/Wildfire.bat
+
+[Details]
+# These values are optional but useful metadata for helper programs
+# Name of the bot's creator/developer
+developer = r0bbi3
+
+# Short description of the bot
+description = Hyper-aggression meets RLBot.
+
+# Fun fact about the bot
+fun_fact = Hates playing defence.
+
+# Link to github repository
+github = https://github.com/robbai/Wildfire
+
+# Programming language
+language = java
