+++ RLBotPack/Wildfire_League_S0W5/python/wildfireAppearance.cfg
@@ -1,0 +1,56 @@
+[Bot Loadout]
+# Primary Color selection
+team_color_id = 35
+# Secondary Color selection
+custom_color_id = 0
+# Car type (Octane, Merc, etc
+car_id = 600
+# Type of decal
+decal_id = 0
+# Wheel selection
+wheels_id = 376
+# Boost selection
+boost_id = 63
+# Antenna Selection
+antenna_id = 0
+# Hat Selection
+hat_id = 0
+# Paint Type (for first color)
+paint_finish_id = 270
+# Paint Type (for secondary color)
+custom_finish_id = 270
+# Engine Audio Selection
+engine_audio_id = 0
+# Car trail Selection
+trails_id = 1948
+# Goal Explosion Selection
+goal_explosion_id = 1903
+
+[Bot Loadout Orange]
+# Primary Color selection
+team_color_id = 33
+# Secondary Color selection
+custom_color_id = 0
+# Car type (Octane, Merc, etc
+car_id = 600
+# Type of decal
+decal_id = 0
+# Wheel selection
+wheels_id = 376
+# Boost selection
+boost_id = 63
+# Antenna Selection
+antenna_id = 0
+# Hat Selection
+hat_id = 0
+# Paint Type (for first color)
+paint_finish_id = 270
+# Paint Type (for secondary color)
+custom_finish_id = 270
+# Engine Audio Selection
+engine_audio_id = 0
+# Car trail Selection
+trails_id = 1948
+# Goal Explosion Selection
+goal_explosion_id = 1903
+
