+++ RLBotPack/Wildfire_League_S0W5/version.txt
@@ -1,0 +1,1 @@
+This version of Wildfire that played in Week 5 of the first (test) season of League Play!
