--- RLBotPack/Wildfire_Wintertide_v0.25/python/README_Tournament.md
@@ -1,19 +1,0 @@
-These instructions are intended for a tournament organizer. In theory, you're reading this because
-you've just extracted a zip file submitted to a tournament. If that's not your situation, go look at README.md
-instead.
-
-1. Install Java 8 or newer.
-2. Reference the included `.cfg` file as you would do with a plain python bot.
-3. Before running the match, double click on the `.bat` file included in
-this submission. It should stay open and manage the java bot.
-
-
-You only need to run a single instance of the `.bat`, and it can handle all the bots in this zip
-in the whole match, even on opposite teams.
-
-
-Advanced:
-
-- It's fine to close and restart `.bat` while the framework is active.
-- If there is a port conflict, you can modify both copies of `port.cfg`
-to use a different one. There's one here and one in the bin folder.
