--- RLBotPack/Wildfire_Wintertide_v0.25/python/wildfire.cfg
@@ -1,29 +1,0 @@
-[Locations]
-# Path to loadout config from runner
-looks_config = wildfireAppearance.cfg
-
-# Path to module from runner
-python_file = wildfire.py
-
-# Name that will be displayed in game
-name = Wildfire
-
-[Bot Parameters]
-java_executable_path = ../Wildfire/bin/Wildfire.bat
-
-[Details]
-# These values are optional but useful metadata for helper programs
-# Name of the bot's creator/developer
-developer = r0bbi3
-
-# Short description of the bot
-description = Full of borderline unexplainable numbers that work together in complete unison - a wonderful mess.
-
-# Fun fact about the bot
-fun_fact = Just wishes it could be recognised as a strong bot; it doesn't fight because it wants to, it fights because that's all it's been taught to do.
-
-# Link to github repository
-github = 
-
-# Programming language
-language = java
