--- RLBotPack/Wildfire_Wintertide_v0.25/version.txt
@@ -1,1 +1,0 @@
-This version of Wildfire that was intended to play in the Wintertide tournament!
